+++ cmd/gen-commands.ts
@@ -299,6 +299,7 @@ export const commands = [
     "CmdConfigurePageGrid", "Configure Page Grid...",
     "CmdToggleEditPDF", "Toggle Edit PDF",
     "CmdApplyRedactions", "Apply Redactions",
+    "CmdCutAnnotation", "Cut Annotation",
     "CmdCopyAnnotation", "Copy Annotation",
     "CmdPasteAnnotation", "Paste Annotation",
     "CmdNone", "Do nothing",